[refactor/ACP-119] Android: ViewModel로 DB 로직 분리
